@@ -1,6 +1,6 @@
 {
   "name": "legendarydrew/laravel-siteimage",
-  "version": "1.0.10",
+  "version": "1.0.11",
   "description": "Laravel package for local environment friendly managing of cloud-hosted images.",
   "keywords": [
     "legendarydrew",
@@ -18,7 +18,7 @@
     }
   ],
   "require": {
-    "php": ">=7.3|^8.0",
+    "php": "^8.2",
     "ext-json": "*",
     "cloudinary/cloudinary_php": "^2.6",
     "intervention/image": "^2.7",
@@ -26,8 +26,8 @@
   },
   "require-dev": {
     "fakerphp/faker": "^1.18",
-    "orchestra/testbench": "^6.0",
-    "phpunit/phpunit": "^9.0"
+    "orchestra/testbench": "^9.7",
+    "phpunit/phpunit": "^10"
   },
   "autoload": {
     "psr-4": {
@@ -41,7 +41,13 @@
   },
   "scripts": {
     "test": "vendor/bin/phpunit tests",
-    "test-coverage": "vendor/bin/phpunit tests --coverage-html coverage"
+    "test-coverage": "vendor/bin/phpunit tests --coverage-html coverage",
+    "post-autoload-dump": [
+      "Illuminate\\Foundation\\ComposerScripts::postAutoloadDump"
+    ],
+    "post-update-cmd": [
+      "Illuminate\\Foundation\\ComposerScripts::postUpdate"
+    ]
   },
   "config": {
     "sort-packages": true
